feat(term): Add reverse function for data manipulation (#24)
Introduces a new built-in function, `reverse`, for manipulating the byte
representation of terms. This provides a convenient way to handle byte
order and data reversal directly within protocol rules. A primary use
case is changing the endianness of integer terms from the default Little
Endian to Big Endian.

The function signature is reverse(data):
- `data`: The term to be reversed. It can be a constant of type `int`,
  `string`, or `[]byte`.